Reorganized namespaces in integration test project.
