+++ Fabric.Identity.IntegrationTests/ControllerTests/CouchDb/CouchDbClientRegistrationTests.cs
@@ -1,6 +1,7 @@
 ﻿using Fabric.Identity.API;
+using Fabric.Identity.IntegrationTests.ControllerTests.InMemory;
 
-namespace Fabric.Identity.IntegrationTests
+namespace Fabric.Identity.IntegrationTests.ControllerTests.CouchDb
 {
     public class CouchDbClientRegistrationTests : ClientRegistrationTests
     {
